discord_bot/pyproject.toml
[tool.poetry]
name = "discord_bot"
version = "0.1.0"
description = ""
authors = ["bafaurazan <rafalbe777@gmail.com>"]
readme = "README.md"
packages = [{include = "bot_config"}]


[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
discord = "^2.3.2"
gtts = "^2.5.1"
black = "^24.4.0"
ruff = "^0.4.1"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
